Ticket Edit confirmation, various small changes
Put the confirmation pop-up that Brendan made into the edit form, made sense to ask to confirm changes. Added the unique toast messages for when any of the fields are empty, so the user knows to fill them in. Noticed that the field names in 'TicketList' were labeled differently than in 'FileForm' ('Part' vs 'Part #') so I changed them to the same for consistency.